[pathing] made pixel references floats, moved Graph constructpr

+++ src/EntityCircle.h
@@ -41,18 +41,18 @@ public:
 
     void update_colour();
     EntityCircle* set_position_randomly(
-        const int min_x,
-        const int max_x,
-        const int min_y,
-        const int max_y
+        const float min_x,
+        const float max_x,
+        const float min_y,
+        const float max_y
     );
     void set_position_to_home();
     EntityCircle* move_to_links(const int offset = 100);
     void set_position_relative_to_links(
-        const int min_x,
-        const int max_x,
-        const int min_y,
-        const int max_y,
+        const float min_x,
+        const float max_x,
+        const float min_y,
+        const float max_y,
         const int offset = 100
     );
     void move_to_destination(
